Baron et al. (2016) Human Pancreas Single-Cell RNA-Seq Reference Data

A SingleCellExperiment object containing human pancreas single-cell RNA-seq data from Baron et al. (2016). This dataset serves as the reference dataset in the scLearn package for cell type annotation.

Usage

RefCellData

Format

A SingleCellExperiment object with 20125 genes (rows) and 8569 cells (columns). The object contains:

counts

Raw count matrix

logcounts

Log-normalized expression values

colData

Cell metadata including:

cell_type1: Cell type annotations

other metadata fields...

Source

https://doi.org/10.1016/j.cels.2016.08.011

References

Baron, M., Veres, A., Wolock, S. L., Faust, A. L., Gaujoux, R., Vetere, A., ... & Melton, D. A. (2016). A single-cell transcriptomic map of the human and mouse pancreas reveals inter- and intra-cell population structure. Cell systems, 3(4), 346-360.

Examples

library(SingleCellExperiment)
data(RefCellData)
# Access expression matrix
assay(RefCellData, "logcounts")[1:5,1:5]
#>        human3_lib2.final_cell_0129 human3_lib2.final_cell_0359
#> A1BG                             0                           0
#> A1CF                             0                           0
#> A2M                              0                           0
#> A2ML1                            0                           0
#> A4GALT                           0                           0
#>        human3_lib1.final_cell_0076 human3_lib3.final_cell_0010
#> A1BG                             0                           0
#> A1CF                             0                           0
#> A2M                              0                           0
#> A2ML1                            0                           0
#> A4GALT                           0                           0
#>        human3_lib1.final_cell_0096
#> A1BG                             0
#> A1CF                             0
#> A2M                              0
#> A2ML1                            0
#> A4GALT                           0
# Access cell metadata
head(colData(RefCellData))
#> DataFrame with 6 rows and 30 columns
#>                                 human cell_type1 total_features
#>                             <numeric>   <factor>      <integer>
#> human3_lib2.final_cell_0129         3     acinar           1974
#> human3_lib2.final_cell_0359         3     acinar           1571
#> human3_lib1.final_cell_0076         3     acinar           2699
#> human3_lib3.final_cell_0010         3     acinar           2648
#> human3_lib1.final_cell_0096         3     acinar           2728
#> human3_lib4.final_cell_0673         3     acinar            870
#>                             log10_total_features total_counts
#>                                        <numeric>    <integer>
#> human3_lib2.final_cell_0129              3.29557        11004
#> human3_lib2.final_cell_0359              3.19645         9013
#> human3_lib1.final_cell_0076              3.43136        12330
#> human3_lib3.final_cell_0010              3.42308        17228
#> human3_lib1.final_cell_0096              3.43600        13041
#> human3_lib4.final_cell_0673              2.94002         3178
#>                             log10_total_counts pct_counts_top_50_features
#>                                      <numeric>                  <numeric>
#> human3_lib2.final_cell_0129            4.04159                    56.3977
#> human3_lib2.final_cell_0359            3.95492                    59.0259
#> human3_lib1.final_cell_0076            4.09100                    43.6172
#> human3_lib3.final_cell_0010            4.23626                    56.9886
#> human3_lib1.final_cell_0096            4.11534                    44.1147
#> human3_lib4.final_cell_0673            3.50229                    50.4405
#>                             pct_counts_top_100_features
#>                                               <numeric>
#> human3_lib2.final_cell_0129                     66.2577
#> human3_lib2.final_cell_0359                     70.4427
#> human3_lib1.final_cell_0076                     54.7040
#> human3_lib3.final_cell_0010                     65.8695
#> human3_lib1.final_cell_0096                     55.4482
#> human3_lib4.final_cell_0673                     63.1215
#>                             pct_counts_top_200_features
#>                                               <numeric>
#> human3_lib2.final_cell_0129                     75.0091
#> human3_lib2.final_cell_0359                     79.1967
#> human3_lib1.final_cell_0076                     65.6285
#> human3_lib3.final_cell_0010                     73.9784
#> human3_lib1.final_cell_0096                     66.0532
#> human3_lib4.final_cell_0673                     75.2045
#>                             pct_counts_top_500_features
#>                                               <numeric>
#> human3_lib2.final_cell_0129                     84.0876
#> human3_lib2.final_cell_0359                     87.5846
#> human3_lib1.final_cell_0076                     76.3747
#> human3_lib3.final_cell_0010                     82.6503
#> human3_lib1.final_cell_0096                     76.7733
#> human3_lib4.final_cell_0673                     88.3575
#>                             total_features_endogenous
#>                                             <integer>
#> human3_lib2.final_cell_0129                      1974
#> human3_lib2.final_cell_0359                      1571
#> human3_lib1.final_cell_0076                      2699
#> human3_lib3.final_cell_0010                      2648
#> human3_lib1.final_cell_0096                      2728
#> human3_lib4.final_cell_0673                       870
#>                             log10_total_features_endogenous
#>                                                   <numeric>
#> human3_lib2.final_cell_0129                         3.29557
#> human3_lib2.final_cell_0359                         3.19645
#> human3_lib1.final_cell_0076                         3.43136
#> human3_lib3.final_cell_0010                         3.42308
#> human3_lib1.final_cell_0096                         3.43600
#> human3_lib4.final_cell_0673                         2.94002
#>                             total_counts_endogenous
#>                                           <integer>
#> human3_lib2.final_cell_0129                   11004
#> human3_lib2.final_cell_0359                    9013
#> human3_lib1.final_cell_0076                   12330
#> human3_lib3.final_cell_0010                   17228
#> human3_lib1.final_cell_0096                   13041
#> human3_lib4.final_cell_0673                    3178
#>                             log10_total_counts_endogenous pct_counts_endogenous
#>                                                 <numeric>             <numeric>
#> human3_lib2.final_cell_0129                       4.04159                   100
#> human3_lib2.final_cell_0359                       3.95492                   100
#> human3_lib1.final_cell_0076                       4.09100                   100
#> human3_lib3.final_cell_0010                       4.23626                   100
#> human3_lib1.final_cell_0096                       4.11534                   100
#> human3_lib4.final_cell_0673                       3.50229                   100
#>                             pct_counts_top_50_features_endogenous
#>                                                         <numeric>
#> human3_lib2.final_cell_0129                               56.3977
#> human3_lib2.final_cell_0359                               59.0259
#> human3_lib1.final_cell_0076                               43.6172
#> human3_lib3.final_cell_0010                               56.9886
#> human3_lib1.final_cell_0096                               44.1147
#> human3_lib4.final_cell_0673                               50.4405
#>                             pct_counts_top_100_features_endogenous
#>                                                          <numeric>
#> human3_lib2.final_cell_0129                                66.2577
#> human3_lib2.final_cell_0359                                70.4427
#> human3_lib1.final_cell_0076                                54.7040
#> human3_lib3.final_cell_0010                                65.8695
#> human3_lib1.final_cell_0096                                55.4482
#> human3_lib4.final_cell_0673                                63.1215
#>                             pct_counts_top_200_features_endogenous
#>                                                          <numeric>
#> human3_lib2.final_cell_0129                                75.0091
#> human3_lib2.final_cell_0359                                79.1967
#> human3_lib1.final_cell_0076                                65.6285
#> human3_lib3.final_cell_0010                                73.9784
#> human3_lib1.final_cell_0096                                66.0532
#> human3_lib4.final_cell_0673                                75.2045
#>                             pct_counts_top_500_features_endogenous
#>                                                          <numeric>
#> human3_lib2.final_cell_0129                                84.0876
#> human3_lib2.final_cell_0359                                87.5846
#> human3_lib1.final_cell_0076                                76.3747
#> human3_lib3.final_cell_0010                                82.6503
#> human3_lib1.final_cell_0096                                76.7733
#> human3_lib4.final_cell_0673                                88.3575
#>                             total_features_feature_control
#>                                                  <integer>
#> human3_lib2.final_cell_0129                              0
#> human3_lib2.final_cell_0359                              0
#> human3_lib1.final_cell_0076                              0
#> human3_lib3.final_cell_0010                              0
#> human3_lib1.final_cell_0096                              0
#> human3_lib4.final_cell_0673                              0
#>                             log10_total_features_feature_control
#>                                                        <numeric>
#> human3_lib2.final_cell_0129                                    0
#> human3_lib2.final_cell_0359                                    0
#> human3_lib1.final_cell_0076                                    0
#> human3_lib3.final_cell_0010                                    0
#> human3_lib1.final_cell_0096                                    0
#> human3_lib4.final_cell_0673                                    0
#>                             total_counts_feature_control
#>                                                <integer>
#> human3_lib2.final_cell_0129                            0
#> human3_lib2.final_cell_0359                            0
#> human3_lib1.final_cell_0076                            0
#> human3_lib3.final_cell_0010                            0
#> human3_lib1.final_cell_0096                            0
#> human3_lib4.final_cell_0673                            0
#>                             log10_total_counts_feature_control
#>                                                      <numeric>
#> human3_lib2.final_cell_0129                                  0
#> human3_lib2.final_cell_0359                                  0
#> human3_lib1.final_cell_0076                                  0
#> human3_lib3.final_cell_0010                                  0
#> human3_lib1.final_cell_0096                                  0
#> human3_lib4.final_cell_0673                                  0
#>                             pct_counts_feature_control total_features_ERCC
#>                                              <numeric>           <integer>
#> human3_lib2.final_cell_0129                          0                   0
#> human3_lib2.final_cell_0359                          0                   0
#> human3_lib1.final_cell_0076                          0                   0
#> human3_lib3.final_cell_0010                          0                   0
#> human3_lib1.final_cell_0096                          0                   0
#> human3_lib4.final_cell_0673                          0                   0
#>                             log10_total_features_ERCC total_counts_ERCC
#>                                             <numeric>         <integer>
#> human3_lib2.final_cell_0129                         0                 0
#> human3_lib2.final_cell_0359                         0                 0
#> human3_lib1.final_cell_0076                         0                 0
#> human3_lib3.final_cell_0010                         0                 0
#> human3_lib1.final_cell_0096                         0                 0
#> human3_lib4.final_cell_0673                         0                 0
#>                             log10_total_counts_ERCC pct_counts_ERCC
#>                                           <numeric>       <numeric>
#> human3_lib2.final_cell_0129                       0               0
#> human3_lib2.final_cell_0359                       0               0
#> human3_lib1.final_cell_0076                       0               0
#> human3_lib3.final_cell_0010                       0               0
#> human3_lib1.final_cell_0096                       0               0
#> human3_lib4.final_cell_0673                       0               0
#>                             is_cell_control
#>                                   <logical>
#> human3_lib2.final_cell_0129           FALSE
#> human3_lib2.final_cell_0359           FALSE
#> human3_lib1.final_cell_0076           FALSE
#> human3_lib3.final_cell_0010           FALSE
#> human3_lib1.final_cell_0096           FALSE
#> human3_lib4.final_cell_0673           FALSE